Remove intermediate library, centralize externs
