+++ src/middleware/mod.rs
@@ -1,5 +1,5 @@
-mod logging;
-pub use self::logging::LogRequest;
-
 mod headers;
+mod logging;
+
 pub use self::headers::CommonHeaders;
+pub use self::logging::LogRequest;
