+++ src/store_limits.rs
@@ -1,6 +1,6 @@
-use super::redis::Script;
-use super::error::BrokerResult;
-use super::store::Store;
+use error::BrokerResult;
+use redis::Script;
+use store::Store;
 
 
 /// Represents a ratelimit
